+++ modules/apps/foundation/contacts/contacts-uad/src/main/java/com/liferay/contacts/uad/exporter/BaseEntryUADExporter.java
@@ -1,0 +1,91 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.contacts.uad.exporter;
+
+import com.liferay.contacts.model.Entry;
+import com.liferay.contacts.service.EntryLocalService;
+import com.liferay.contacts.uad.constants.ContactsUADConstants;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.util.StringBundler;
+
+import com.liferay.user.associated.data.exporter.DynamicQueryUADExporter;
+
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * Provides the base implementation for the entry UAD exporter.
+ *
+ * <p>
+ * This implementation exists only as a container for the default methods
+ * generated by ServiceBuilder. All custom service methods should be put in
+ * {@link EntryUADExporter}.
+ * </p>
+ *
+ * @author Brian Wing Shun Chan
+ * @generated
+ */
+public abstract class BaseEntryUADExporter extends DynamicQueryUADExporter<Entry> {
+	@Override
+	protected ActionableDynamicQuery doGetActionableDynamicQuery() {
+		return entryLocalService.getActionableDynamicQuery();
+	}
+
+	@Override
+	protected String[] doGetUserIdFieldNames() {
+		return ContactsUADConstants.USER_ID_FIELD_NAMES_ENTRY;
+	}
+
+	@Override
+	protected String toXmlString(Entry entry) {
+		StringBundler sb = new StringBundler(22);
+
+		sb.append("<model><model-name>");
+		sb.append("com.liferay.contacts.model.Entry");
+		sb.append("</model-name>");
+
+		sb.append(
+			"<column><column-name>entryId</column-name><column-value><![CDATA[");
+		sb.append(entry.getEntryId());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>userId</column-name><column-value><![CDATA[");
+		sb.append(entry.getUserId());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>userName</column-name><column-value><![CDATA[");
+		sb.append(entry.getUserName());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>fullName</column-name><column-value><![CDATA[");
+		sb.append(entry.getFullName());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>emailAddress</column-name><column-value><![CDATA[");
+		sb.append(entry.getEmailAddress());
+		sb.append("]]></column-value></column>");
+		sb.append(
+			"<column><column-name>comments</column-name><column-value><![CDATA[");
+		sb.append(entry.getComments());
+		sb.append("]]></column-value></column>");
+
+		sb.append("</model>");
+
+		return sb.toString();
+	}
+
+	@Reference
+	protected EntryLocalService entryLocalService;
+}
